🚚refactor: img, icon 구조 변경 및 import 수정
