--- docs/check.md
@@ -1,21 +1,0 @@
-## 1. 체크리스트
-
-### 랜딩 페이지
-
-- [] 로컬 스토리지에 accessToken이 있는 경우 상단바 ‘로그인’ 버튼이 판다 이미지로 바뀝니다.
-
-### 로그인
-
-- [x] 회원가입을 성공한 정보를 입력하고 스웨거 명세된 “/auth/signIn”으로 POST 요청을 하면 로그인이 완료됩니다.
-- [] 로그인이 완료되면 로컬 스토리지에 accessToken을 저장하고 “/” 로 이동합니다.
-- [] 로그인/회원가입 페이지에 접근시 로컬 스토리지에 accessToken이 있는 경우 ‘/’ 페이지로 이동합니다.
-
-### 회원가입
-
-- [x] 유효한 정보를 입력하고 스웨거 명세된 “/auth/signUp”으로 POST 요청해서 성공 응답을 받으면 회원가입이 완료됩니다.
-- [] 회원가입이 완료되면 “/login”로 이동합니다.
-- [] 회원가입 페이지에 접근시 로컬 스토리지에 accessToken이 있는 경우 ‘/’ 페이지로 이동합니다.
-
-### 심화 요구사항
-
-- [] 로그인, 회원가입 기능에 react-hook-form을 활용해봅니다.
